# docs/swagger.yaml
basePath: /home-work-z-api
definitions:
  vo.OrderPutVO:
    properties:
      buy:
        example: false
        type: boolean
      date:
        example: 訂單日期(yyyy-MM-dd)
        type: string
      limit_price:
        example: 0
        type: number
      market_price:
        example: 0
        type: number
      quantity:
        example: 0
        type: integer
      sell:
        example: true
        type: boolean
    required:
    - date
    - quantity
    type: object
  vo.OrderVO:
    properties:
      buy:
        type: boolean
      date:
        type: string
      id:
        type: integer
      limit_price:
        type: number
      market_price:
        type: number
      quantity:
        type: integer
      sell:
        type: boolean
    type: object
  vo.UserPostVO:
    properties:
      account:
        default: alan
        example: 帳號
        type: string
      password:
        default: home_work_z
        example: 密碼
        type: string
    required:
    - account
    - password
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: qoo753951ooq@gmail.com
    name: Alan Syu
  description: Home work Z Swagger 說明文件
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Home Work Z API
  version: "1.0"
paths:
  /order:
    get:
      description: 訂單列表
      parameters:
      - default: "2001-01-01"
        description: 開始時間 (yyyy-MM-dd)
        in: query
        name: starttime
        type: string
      - default: "2001-01-02"
        description: 結束時間 (yyyy-MM-dd)
        in: query
        name: endtime
        type: string
      - default: 10
        description: 取前幾筆
        in: query
        name: top
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vo.OrderVO'
            type: array
      security:
      - BearerAuth: []
      summary: getList
      tags:
      - trade
    post:
      consumes:
      - multipart/form-data
      description: 新增訂單
      parameters:
      - default: false
        in: formData
        name: buy
        required: true
        type: boolean
      - default: 訂單日期(yyyy-MM-dd)
        in: formData
        name: date
        required: true
        type: string
      - default: 0
        in: formData
        name: limit_price
        type: number
      - default: 0
        in: formData
        name: market_price
        type: number
      - default: 0
        in: formData
        name: quantity
        required: true
        type: integer
      - default: false
        in: formData
        name: sell
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.OrderVO'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: addOne
      tags:
      - trade
  /order/{id}:
    delete:
      description: 刪除訂單
      parameters:
      - description: 編號
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: deleteOne
      tags:
      - trade
    get:
      description: 單筆訂單
      parameters:
      - description: 編號
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.OrderVO'
      security:
      - BearerAuth: []
      summary: getOne
      tags:
      - trade
    put:
      consumes:
      - application/json
      description: 編輯訂單
      parameters:
      - description: 編號
        in: path
        name: id
        required: true
        type: integer
      - description: body for OrderPutVO content
        in: body
        name: putVO
        required: true
        schema:
          $ref: '#/definitions/vo.OrderPutVO'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: editOne
      tags:
      - trade
  /user/login:
    post:
      consumes:
      - application/json
      description: 登入
      parameters:
      - description: body for UserPostVO content
        in: body
        name: postVO
        required: true
        schema:
          $ref: '#/definitions/vo.UserPostVO'
      produces:
      - text/plain
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: login
      tags:
      - authorization
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"